+++ Wasmachine/inc/WasProgrammaUitvoerenController.h
@@ -24,7 +24,7 @@ private:
     bool m_heaterOn = false;
     bool m_waterValveOpen = false;
     bool m_pumpOn = false;
-    int m_currentRPM = 0;
+    uint8_t m_currentRPM = 0;
     
     bool m_perIOHandSuspend = true;
